+++ Sources/Domain/CoreDomain/use_cases/GetHero.swift
@@ -1,5 +1,6 @@
 import CoreDomainContracts
 
+@MainActor
 public struct GetHero: GetHeroProtocol
 {
     // MARK: Dependencies