+++ Sources/Features/HeroList/ListHeroesPresenter.swift
@@ -1,45 +1,59 @@
 import CoreDomainContracts
 
-public protocol ListHeroesPresenterProtocol: AnyObject {
+/// This protocol is @MainActor-isolated because of a Swift bug. See `MarvelRepository` for more details.
+///
+@MainActor
+public protocol ListHeroesPresenterProtocol: AnyObject
+{
     var ui: ListHeroesUI? { get set }
     func screenTitle() -> String
     func getHeroes()
 }
 
+/// This protocol is @MainActor-isolated because of a Swift bug. See `MarvelRepository` for more details.
+///
 @MainActor
-public protocol ListHeroesUI: AnyObject, Sendable {
+public protocol ListHeroesUI: AnyObject, Sendable
+{
     func update(heroes: [CharacterModel])
 }
 
-public final class ListHeroesPresenter: ListHeroesPresenterProtocol {
+@MainActor
+public final class ListHeroesPresenter: ListHeroesPresenterProtocol
+{
     public weak var ui: ListHeroesUI?
-    private let getHeroesUseCase: GetHeroesProtocol
+    private let fetchHeroesUseCase: FetchNextPageProtocol
 
-    public static func make(getHeroes: GetHeroesProtocol) -> ListHeroesPresenterProtocol {
-        ListHeroesPresenter(getHeroesUseCase: getHeroes)
+    public static func make(
+        fetchHeroes: FetchNextPageProtocol
+    ) -> ListHeroesPresenterProtocol {
+        ListHeroesPresenter(fetchHeroesUseCase: fetchHeroes)
     }
 
-    init(getHeroesUseCase: GetHeroesProtocol) {
-        self.getHeroesUseCase = getHeroesUseCase
+    init(fetchHeroesUseCase: FetchNextPageProtocol)
+    {
+        self.fetchHeroesUseCase = fetchHeroesUseCase
     }
-    
+
     public func screenTitle() -> String {
         "List of Heroes"
     }
-    
+
     // MARK: UseCases
-    
+
     public func getHeroes() {
         guard let ui else {
             assertionFailure("ListHeroesPresenter.ui is nil")
             return
         }
-        getHeroesUseCase { characterModelContainer in
-            print("Characters \(characterModelContainer.results)")
-            Task { @MainActor in
-                ui.update(heroes: characterModelContainer.results)
-            }
-        }
+        fetchHeroesUseCase(
+            onSuccess: { characterModelContainers in
+                Task { @MainActor in
+                    ui.update(heroes: characterModelContainers.flatMap { container in container.results })
+                }
+            },
+            onFailure: { _ in }
+        )
     }
 }