+++ Tests/Domain/CoreDomainTests/repositories/MarvelRepositoryTests.swift
@@ -5,13 +5,15 @@ import CoreDomainContractsTestSupport
 import Testing
 import Foundation
 
-@Suite
-struct MarvelRepositoryTests {
-    
+/// This struct is @MainActor-isolated because of a Swift bug. See `MarvelRepository` for more details.
+/// 
+@Suite @MainActor
+struct MarvelRepositoryTests
+{
     @Test("When fetching heroes, it retrieves data from remote and stores it locally")
-    func fetchHeroesSuccessfully() async {
+    func fetchHeroesSuccessfully() async throws {
         // Given
-        let localStore = await CharacterModelContainerStoreProtocolMock()
+        let localStore = CharacterModelContainerStore(pages: [], nextOffset: 0, lockedOffsets: [])
         let remote = MarvelRemoteDataSourceProtocolMock()
         let repository = MarvelRepository(local: localStore, remote: remote)
         let characterDataContainer = CharacterDataContainer(
@@ -29,23 +31,28 @@ struct MarvelRepositoryTests {
                 ]
             )
         )
-        remote.fetchHeroesCompletionSendableEscapingCharacterDataContainerVoidVoidClosure = { completion in
-            completion(characterDataContainer)
+        remote.fetchHeroesLimitIntOffsetIntOnSuccessSendableEscapingCharacterDataContainerVoidOnFailureSendableEscapingErrorVoidVoidClosure = { limit, offset, onSuccess, onFailure in
+            onSuccess(characterDataContainer)
         }
-        
+
         // When
-        let result = await withCheckedContinuation { continuation in
-            repository.fetchHeroes { container in
-                continuation.resume(returning: container)
-            }
+        let result = try await withCheckedThrowingContinuation { continuation in
+            repository.fetchHeroes(
+                onSuccess: { container in
+                    continuation.resume(returning: container)
+                },
+                onFailure: { error in
+                    continuation.resume(throwing: error)
+                }
+            )
         }
         
         // Then
         #expect(result.count == 1)
-        #expect(result.results.first?.id == 1)
-        #expect(result.results.first?.name == "Spider-Man")
-        #expect(result.results.first?.thumbnail?.absoluteString == "http://example.com/image/portrait_small.jpg")
-        #expect(await localStore.addPageCharacterModelContainerVoidCallsCount == 1)
+        #expect(result[0].count == 1)
+        #expect(result[0].results.first?.id == 1)
+        #expect(result[0].results.first?.name == "Spider-Man")
+        #expect(result[0].results.first?.thumbnail?.absoluteString == "http://example.com/image/portrait_small.jpg")
     }
     
     @Test("When retrieving a hero by ID, it fetches from local storage")